H13-14Ogrenci_Perfonmans_Tahminlemesi_MLOps-main/requirements.txt
setuptools
python-box
ensure
PyYAML
joblib
scikit-learn
pickle4
pandas
catboost
